Refactoring usage of smart pointer on Matrix, Perceptron and MLP implementations

+++ NeuralNetwork/Common/ActivationFunctions.h
@@ -2,10 +2,10 @@
 #define AIENGINE_ACTIVATIONFUNCTIONS_H
 #include <cmath>
 
-float sigmoid(float value);
-float sign(float value);
-float zeroOrOne(float value);
-float noActivation(float value);
-void setZeroOrOneThreshold(float threshold);
+double sigmoid(double value);
+double sign(double value);
+double zeroOrOne(double value);
+double noActivation(double value);
+void setZeroOrOneThreshold(double threshold);
 
 #endif //AIENGINE_ACTIVATIONFUNCTIONS_H
